move publish with descendants
