+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/modals/constants.ts
@@ -1,3 +1,2 @@
-export * from './publish-with-descendants-modal/constants.js';
 export * from './save-modal/constants.js';
 export * from './document-picker-modal.token.js';
