+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/modals/manifests.ts
@@ -1,4 +1,3 @@
-import { manifest as publishWithDescendantsModalManifest } from './publish-with-descendants-modal/manifest.js';
 import { manifest as saveModalManifest } from './save-modal/manifest.js';
 
-export const manifests: Array<UmbExtensionManifest> = [publishWithDescendantsModalManifest, saveModalManifest];
+export const manifests: Array<UmbExtensionManifest> = [saveModalManifest];
