--- src/Umbraco.Web.UI.Client/src/packages/documents/documents/modals/publish-with-descendants-modal/constants.ts
@@ -1,2 +1,0 @@
-export * from './document-publish-with-descendants-modal.token.js';
-export { UMB_DOCUMENT_PUBLISH_WITH_DESCENDANTS_MODAL_ALIAS } from './manifest.js';
