+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/modals/types.ts
@@ -4,7 +4,5 @@ import type { UmbContentVariantPickerData, UmbContentVariantPickerValue } from '
 export type UmbDocumentVariantPickerData = UmbContentVariantPickerData<UmbDocumentVariantOptionModel>;
 export type UmbDocumentVariantPickerValue = UmbContentVariantPickerValue;
 
-export type * from '../publishing/publish/modal/constants.js';
-export type * from './publish-with-descendants-modal/constants.js';
 export type * from './save-modal/constants.js';
 export type * from './document-picker-modal.token.js';
