+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/constants.ts
@@ -2,3 +2,4 @@ export * from './publish/constants.js';
 export * from './repository/constants.js';
 export * from './schedule-publish/constants.js';
 export * from './unpublish/constants.js';
+export * from './publish-with-descendants/constants.js';
