+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/manifests.ts
@@ -3,6 +3,7 @@ import { manifests as publishManifest } from './publish/manifests.js';
 import { manifests as repositoryManifests } from './repository/manifests.js';
 import { manifests as schedulePublishManifests } from './schedule-publish/manifests.js';
 import { manifests as unpublishManifests } from './unpublish/manifests.js';
+import { manifests as publishWithDescendantsManifest } from './publish-with-descendants/manifests.js';
 import { UMB_WORKSPACE_CONDITION_ALIAS } from '@umbraco-cms/backoffice/workspace';
 
 export const manifests: Array<UmbExtensionManifest> = [
@@ -19,6 +20,7 @@ export const manifests: Array<UmbExtensionManifest> = [
 		],
 	},
 	...publishManifest,
+	...publishWithDescendantsManifest,
 	...repositoryManifests,
 	...schedulePublishManifests,
 	...unpublishManifests,
