+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/publish-with-descendants/modal/document-publish-with-descendants-modal.element.ts
@@ -1,5 +1,3 @@
-import { UmbDocumentVariantState, type UmbDocumentVariantOptionModel } from '../../types.js';
-import { isNotPublishedMandatory } from '../utils.js';
 import type {
 	UmbDocumentPublishWithDescendantsModalData,
 	UmbDocumentPublishWithDescendantsModalValue,
@@ -8,6 +6,8 @@ import { css, customElement, html, state } from '@umbraco-cms/backoffice/externa
 import { UmbModalBaseElement } from '@umbraco-cms/backoffice/modal';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
 import { UmbSelectionManager } from '@umbraco-cms/backoffice/utils';
+import { UmbDocumentVariantState, type UmbDocumentVariantOptionModel } from '../../../types.js';
+import { isNotPublishedMandatory } from '../../utils.js';
 
 import '../shared/document-variant-language-picker.element.js';
