+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/publish-with-descendants/modal/document-publish-with-descendants-modal.token.ts
@@ -1,6 +1,7 @@
-import type { UmbDocumentVariantPickerData, UmbDocumentVariantPickerValue } from '../types.js';
-import { UMB_DOCUMENT_PUBLISH_WITH_DESCENDANTS_MODAL_ALIAS } from './manifest.js';
+import type { UmbDocumentVariantPickerData, UmbDocumentVariantPickerValue } from '../../../types.js';
 import { UmbModalToken } from '@umbraco-cms/backoffice/modal';
+
+export const UMB_DOCUMENT_PUBLISH_WITH_DESCENDANTS_MODAL_ALIAS = 'Umb.Modal.DocumentPublishWithDescendants';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
 export interface UmbDocumentPublishWithDescendantsModalData extends UmbDocumentVariantPickerData {}
