+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/publish-with-descendants/modal/types.ts
@@ -1,0 +1,1 @@
+export * from './document-publish-with-descendants-modal.token.js';
