+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/utils.ts
